fixed huge error amount in CustomGyro because of adding axSquared before calculating changed errors from long to float

+++ CustomGyro/CustomGyro.h
@@ -12,8 +12,8 @@ class CustomGyro
         int16_t gx;
         int16_t gy;
         int16_t gz;
-        long _pitchError;
-        long _rollError;
+        float _pitchError;
+        float _rollError;
     public:
         CustomGyro();
         void begin();
